Moved DateUtil into util/
