+++ src/main/java/app/ui/MainSceneController.java
@@ -39,6 +39,7 @@ import javafx.scene.shape.Line;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 import util.ArrayFunctions;
+import util.DateUtil;
 import app.AppError;
 import app.communication.EventQueues;
 import app.communication.Msg;
@@ -54,7 +55,6 @@ import app.mapping.PointData;
 import app.mapping.RangeData;
 import audio.synth.EvInstrEnum;
 import audio.synth.InstrumentEnum;
-import dataRepo.DateUtil;
 import dataRepo.Sonifiable;
 import dataRepo.SonifiableID;
 import dataRepo.FilterFlag;
